+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { Route, BrowserRouter as Router, Routes } from 'react-router-dom';
+import CardCreate from './components/pages/CardCreate';
 import CardEdit from './components/pages/CardEdit';
-import CardForm from './components/pages/CardForm';
 import CardList from './components/pages/CardList';
 
 function App() {
@@ -8,7 +8,7 @@ function App() {
     <Router>
       <Routes>
         <Route path="/" element={<CardList />} />
-        <Route path="/create" element={<CardForm />} />
+        <Route path="/create" element={<CardCreate />} />
         <Route path="/edit/:id" element={<CardEdit />} />
       </Routes>
     </Router>
